+++ src/builds.hpp
@@ -8,7 +8,5 @@
 
 uint32_t calculateCRC32(const std::string& filename);
 
-// Serializes builds to a JSON config file
-bool SaveBuildMetadata(const std::map<std::string, BuildMetadata>& builds);
 // Loads builds from a JSON config file
 std::map<std::string, BuildMetadata> LoadBuildMetadata();
